# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.systemCreateBody:
    properties:
      name:
        type: string
      startedDate:
        type: string
    required:
    - name
    type: object
  controllers.systemCreatePeerBody:
    properties:
      allowed_ip:
        type: string
      buy_date:
        type: string
      config_end_point:
        type: string
      data_limit:
        type: number
      email:
        type: string
      expire_date:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  gin.H:
    additionalProperties: {}
    type: object
...
  models.PeerInfo:
    properties:
      buyDate:
        type: string
      dataLimit:
        type: number
      expireDate:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      usage:
        type: number
    type: object
  models.SystemInfo:
    properties:
      name:
        type: string
      peers:
        items:
          $ref: '#/definitions/models.PeerInfo'
        type: array
      startedDate:
        type: string
      totalUsage:
        type: number
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a Wireguard API server.
  title: Wireguard API
  version: "1.0"
paths:
...
  /systems:
    post:
      consumes:
      - application/json
      description: Create a new system with the specified name.
      parameters:
      - description: System details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.systemCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: System created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Fields to read body
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new system
      tags:
      - Systems
  /systems/{name}:
    get:
      description: description for SystemShow.
      parameters:
      - description: System name
        in: path
        name: name
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: System information
          schema:
            $ref: '#/definitions/models.SystemInfo'
        "400":
          description: Invalid page number" "Invalid per page number" "Invalid system
            fetching
          schema:
            $ref: '#/definitions/gin.H'
      summary: System Show
      tags:
      - Systems
  /systems/{name}/peers:
    post:
      consumes:
      - application/json
      description: Create a new peer for a system by name.
      parameters:
      - description: System name
        in: path
        name: name
        required: true
        type: string
      - description: Peer details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.systemCreatePeerBody'
      produces:
      - application/json
      responses:
        "200":
          description: Peer created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Fields to read body
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new peer for a system
      tags:
      - Systems
  /systems/{name}/reload:
    post:
      description: Reload a system by name.
      parameters:
      - description: System name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: System reloaded
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reload a system
      tags:
      - Systems
  /systems/{name}/usage:
    get:
      description: Retrieve systems based on usage, sorted by usage.
      parameters:
      - description: System name
        in: path
        name: name
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: System information
          schema:
            $ref: '#/definitions/models.SystemInfo'
        "400":
          description: Invalid page number" "Invalid per page number
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get systems based on usage
      tags:
      - Systems
...
  /test/systems/{name}/peers:
    post:
      consumes:
      - application/json
      description: Create a new peer for a system by name.
      parameters:
      - description: System name
        in: path
        name: name
        required: true
        type: string
      - description: Peer details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.systemCreatePeerBody'
      produces:
      - application/json
      responses:
        "200":
          description: Peer created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Fields to read body
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new peer for a system
      tags:
      - Systems
      - Test
  /test/systems/{name}/reload:
    post:
      description: Reload a system by name.
      parameters:
      - description: System name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: System reloaded
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reload a system
      tags:
      - Test
      - Systems
...
swagger: "2.0"